Rename the strategy to DisplayStringIndexStorageStrategy
